requirements.txt
duckdb>=0.10.0
pyyaml>=6.0
pytest>=7.0
